Rename test tags namespace
