+++ src/test/scala/de/uhh/lt/jst/wsd/SenseFeatureAggregatorTest.scala
@@ -3,7 +3,7 @@ package de.uhh.lt.jst.wsd
 import com.holdenkarau.spark.testing.SharedSparkContext
 import org.scalatest._
 import de.uhh.lt.jst.utils.Const
-import de.uhh.lt.testtags.NeedsMissingFiles
+import de.uhh.lt.testing.tags.NeedsMissingFiles
 
 class SenseFeatureAggregatorTest extends FlatSpec with Matchers with SharedSparkContext {
   "The SenseFeatureAggregatorTest object" should "skip wrong trigrams" in {
