+++ src/test/scala/de/uhh/lt/testing/tags/BrokenTest.scala
@@ -1,0 +1,5 @@
+package de.uhh.lt.testing.tags
+
+import org.scalatest.Tag
+
+object BrokenTest extends Tag("de.uhh.lt.testing.tags.BrokenTest")
