+++ src/test/scala/de/uhh/lt/testing/tags/NeedsMissingFiles.scala
@@ -1,8 +1,8 @@
-package de.uhh.lt.testtags
+package de.uhh.lt.testing.tags
 
 import org.scalatest.Tag
 
-object NeedsMissingFiles extends Tag("de.uhh.lt.testtags.NeedsDownload") {
+object NeedsMissingFiles extends Tag("de.uhh.lt.testing.tags.NeedsDownload") {
   /**
     * For donwload instructison, please see comment in
     * [[de.uhh.lt.jst.utils.Const.PRJ_TEST.WSD_RES]]
