--- src/test/scala/de/uhh/lt/testtags/BrokenTest.scala
@@ -1,5 +1,0 @@
-package de.uhh.lt.testtags
-
-import org.scalatest.Tag
-
-object BrokenTest extends Tag("de.uhh.lt.testtags.BrokenTest")
